requirements.txt
Django==1.7
django-debug-toolbar
djangorestframework
markdown
django-filter